fix(cssRules): add additional try/catch
Fixed a bug where FireFox would throw a security exception when trying to access the css rules from a stylesheet loaded from a 3rd party server (Think CDN). This exception would stop Aurelia CSS animations from running and instead just snap. The update tries to read the rules and if it fails it continues on with the loop, where previously it would terminate the loop and always return a false result.

addition to fix for issue #40